requirements.txt
# Essential dependencies for BLG 480E Project 2
# Web crawler and search system with native implementations

# Core HTTP and web crawling (minimal dependencies)
aiohttp==3.8.4                # Async HTTP client for web crawling
beautifulsoup4==4.12.2        # HTML parsing and content extraction

# Database (lightweight)
aiosqlite==0.19.0             # Async SQLite interface

# CLI and configuration
click==8.1.6                  # CLI framework
rich==13.4.2                  # Rich text and beautiful formatting
pyyaml==6.0.1                 # YAML configuration files
pydantic==2.1.1               # Data validation and settings

# Development and testing
pytest==7.4.0                 # Testing framework
pytest-asyncio==0.21.1        # Async testing support
aioresponses==0.7.4           # Mock HTTP responses for testing

# Note: Search engine and indexing are implemented natively using Python built-ins
# No external search libraries (whoosh, elasticsearch, etc.) are used
# Text processing uses native Python re, collections, math modules
# Crawling implements custom backpressure and deduplication logic